pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[project]
name = "notebook-zang"
version = "0.1.0"
description = "Notebook com Skills customizáveis para análise de dados"
requires-python = ">=3.11"
dependencies = [
    # Web
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.18",

    # HTTP & Dados
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # Banco de dados
    "sqlmodel>=0.0.22",
    "aiosqlite>=0.20.0",
    "greenlet>=3.0.0",

    # IA / LLM
    "google-genai>=1.0.0",

    # Documentos
    "pdfplumber>=0.11.0",
    "pypdf>=4.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
